Update AOT compiler

+++ client/tools/utils/seed/clean.ts
@@ -6,7 +6,7 @@ import Config from '../../config';
 
 /**
  * Cleans the given path(s) using `rimraf`.
- * @param {string or string[]} paths - The path or list of paths to clean.
+ * @param {string | string[]} paths - The path or list of paths to clean.
  */
 export function clean(paths: string|string[]): (done: () => void) => void {
   return done => {
@@ -17,13 +17,13 @@ export function clean(paths: string|string[]): (done: () => void) => void {
       pathsToClean = [<string>paths];
     }
 
-    let promises = pathsToClean.map(p => {
+    const promises = pathsToClean.map(p => {
       return new Promise(resolve => {
         const relativePath: string = relative(Config.PROJECT_ROOT, p);
-        //if (relativePath.startsWith('..')) {
-         // util.log(util.colors.bgRed.white(`Cannot remove files outside the project root but tried "${normalize(p)}"`));
-         // process.exit(1);
-        //} else {
+        if (relativePath.startsWith('..')) {
+          util.log(util.colors.bgRed.white(`Cannot remove files outside the project root but tried "${normalize(p)}"`));
+          process.exit(1);
+        } else {
           rimraf(p, e => {
             if (e) {
               util.log('Clean task failed with', e);
@@ -32,7 +32,7 @@ export function clean(paths: string|string[]): (done: () => void) => void {
             }
             resolve();
           });
-       // }
+        }
       });
     });
     Promise.all(promises).then(() => (done || (() => 1))())
